# docs/apidoc.yaml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: US Army Corps of Engineers Pallid Sturgeon
  license:
    name: MIT
servers:
  - url: http://localhost:8080/psapi
paths:
  /projects:
    get:
      summary: List projects
      operationId: GetProjects
      tags:
        - projects
      responses:
        '200':
          description: Array of projects
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/project"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /seasons:
    get:
      summary: List Seasons
      operationId: GetSeasons
      tags:
        - seasons
      responses:
        '200':
          description: Array of seasons
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/season"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /segments:
    get:
      summary: List segments
      operationId: GetSegments
      tags:
        - segments
      responses:
        '200':
          description: Array of segments
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/segment"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bends:
    get:
      summary: List bends
      operationId: GetBends
      tags:
        - bends
      responses:
        '200':
          description: Array of bends
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/bend"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /upload:
    post:
      tags:
      - "upload"
      summary: "upload all data"
      description: ""
      operationId: "addUpload"
      requestBody:
        description: "Collection of different types of uploads"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/upload"
      responses:
        '201':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error" 
components:
  schemas:
    project:
      type: object
      required:
        - code
        - description
      properties:
        code:
          type: integer
          format: int32
        description:
          type: string
    season:
      type: object
      required:
        - id
        - code
        - description
        - fieldAppFlag
        - ProjectCode
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        description:
          type: string
        fieldAppFlag:
          type: string
        ProjectCode:
          type: integer
          format: int32
    segment:
      type: object
      required:
        - id
        - code
        - type
      properties:
        id:
          type: integer
          format: int32
        code:
          type: integer
          format: int32
        description:
          type: string
        type:
          type: string
        upperRiverMile:
          type: integer
          format: int32
        lowerRiverMile:
          type: integer
          format: int32
        rpma:
          type: integer
          format: int32
    bend:
      type: object
      required:
        - id
        - code
        - bendNumber
        - state
      properties:
        id:
          type: integer
          format: int32
        bendNumber:
          type: integer
          format: int32
        description:
          type: string
        segmentCode:
          type: integer
          format: int32
        upperRiverMile:
          type: integer
          format: int32
        lowerRiverMile:
          type: integer
          format: int32
        state:
          type: integer
          format: int32
    upload:
      type: object
      properties:
        siteUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadSite"
        fishUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadFish"
        telemetryUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadTelemetry"
        searchUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadSearch"
        procedureUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadProcedure"
        supplementalUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadSupplemental"
        moriverUpload:
          type: array
          items:
            $ref: "#/components/schemas/uploadMoriver"
    procedureOut:
      type: object
      required:
        - uploadSessionId
        - uploadBy
      properties:
        uploadSessionId:
          type: integer
          format: int32
        uploadBy:
          type: string
          maxLength: 20
        siteCntFinal:
          type: integer
          format: int32
        mrCntFinal:
          type: integer
          format: int32
        fishCntFinal:
          type: integer
          format: int32
        searchCntFinal:
          type: integer
          format: int32
        suppCntFinal:
          type: integer
          format: int32
        telemetryCntFinal:
          type: integer
          format: int32
        procedureCntFinal:
          type: integer
          format: int32
        noSiteCnt:
          type: integer
          format: int32
        siteMatch:
          type: integer
          format: int32
        noSiteIDMsg:
          type: string
    uploadSite:
      type: object
      required:
        - siteId
        - uploadFilename
      properties:
        siteId:
          type: integer
          format: int32
        siteFid:
          type: string
          maxLength: 50
        siteYear:
          type: integer
          format: int32
        fieldofficeID:
          type: string
          maxLength: 10
        fieldOffice:
          type: string
          maxLength: 255
        projectId:
          type: integer
          format: int32
        project:
          type: string
        segmentId:
          type: integer
          format: int32
        segment:
          type: string
          maxLength: 255
        seasonId:
          type: string
          maxLength: 5
        season:
          type: string
          maxLength: 5
        bend:
          type: integer
          format: int32
        bendrn:
          type: string
          maxLength: 10
        bendRiverMile:
          type: number
        comments:
          type: string
          maxLength: 2000
        uploadFilename:
          type: string
          maxLength: 200
    uploadFish:
      type: object
      required:
        - siteId
        - uploadFilename
      properties:
        siteId:
          type: integer
          format: int32
        fFid:
          type: string
          maxLength: 50
        mrFid:
          type: string
          maxLength: 50
        panelhook:
          type: string
          maxLength: 5
        bait:
          type: string
          maxLength: 1
        species:
          type: string
          maxLength: 4
        length:
          type: number
        weight:
          type: number
        fishcount:
          type: integer
          format: int32
        finCurl:
          type: string
          maxLength: 1
        otolith:
          type: string
          maxLength: 1
        rayspine:
          type: string
          maxLength: 1
        scale:
          type: string
          maxLength: 1
        ftprefix:
          type: string
          maxLength: 3
        ftnum:
          type: string
          maxLength: 7
        ftmr:
          type: string
          maxLength: 1
        comments:
          type: string
          maxLength: 2000
        uploadFilename:
          type: string
          maxLength: 200
    uploadTelemetry:
      type: object
      required:
        - tFid
        - uploadFilename
      properties:
        tFid:
          type: string
          maxLength: 50
        seFid:
          type: string
          maxLength: 199
        bend:
          type: number
        radioTagNum:
          type: integer
          format: int32
        frequencyIdCode:
          type: integer
          format: int32
        captureTime:
          type: string
          maxLength: 9
        captureLatitude:
          type: number
        captureLongitude:
          type: number
        positionConfidence:
          type: number
        macroId:
          type: string
          maxLength: 10
        mesoId:
          type: string
          maxLength: 10
        depth:
          type: number
        temp:
          type: number
        conductivity:
          type: number
        turbidity:
          type: number
        silt:
          type: number
        sand:
          type: number
        gravel:
          type: number
        comments:
          type: string
          maxLength: 20
        uploadFilename:
          type: string
          maxLength: 200
    uploadSearch:
      type: object
      required:
        - siteId
        - uploadFilename
      properties:
        siteId:
          type: integer
          format: int32
        seFid:
          type: string
          maxLength: 199
        dsId:
          type: integer
          format: int32
        siteFid:
          type: string
          maxLength: 199
        searchDate:
          type: string
        recorder:
          type: string
          maxLength: 3
        searchTypeCode:
          type: string
          maxLength: 2
        searchDay:
          type: integer
          format: int32
        startTime:
          type: string
          maxLength: 11
        startLatitude:
          type: number
        startLongitude:
          type: number
        stopTime:
          type: string
          maxLength: 11
        stopLatitud:
          type: number
        stopLongitude:
          type: number
        temp:
          type: number
        conductivity:
          type: number
        uploadFilename:
          type: string
          maxLength: 200
    uploadProcedure:
      type: object
      required:
        - id
        - uploadFilename
      properties:
        id:
          type: integer
          format: int32
        f_fid:
          type: string
          maxLength: 50
        purposeCode:
          type: string
          maxLength: 2
        procedurDate:
          type: string
        procedureStartTime:
          type: string
          maxLength: 11
        procedureEndTime:
          type: string
          maxLength: 11
        procedureBy:
          type: string
          maxLength: 3
        antibioticInjectionInd:
          type: integer
          format: int32
        photoDorsalInd:
          type: integer
          format: int32
        photoVentralInd:
          type: integer
          format: int32
        photoLeftInd:
          type: integer
          format: int32
        oldRadioTagNum:
          type: integer
          format: int32
        oldFrequencyId:
          type: integer
          format: int32
        dstSerialNum:
          type: integer
          format: int32
        dstStartDate:
          type: string
        dstStartTime:
          type: string
        dstReimplantInd:
          type: integer
          format: int32
        newRadioTagNum:
          type: integer
          format: int32
        newFrequencyId:
          type: integer
          format: int32
        sexCode:
          type: string
          maxLength: 1
        bloodSampleInd:
          type: integer
          format: int32
        eggSampleInd:
          type: integer
          format: int32
        comments:
          type: string
          maxLength: 2000
        fishHealthComments:
          type: string
          maxLength: 2000
        evalLocationCode:
          type: string
          maxLength: 2
        spawnCode:
          type: string
          maxLength: 2
        visualReproStatusCode:
          type: string
          maxLength: 4
        ultrasoundReproStatusCode:
          type: string
          maxLength: 4
        expectedSpawnYear:
          type: integer
          format: int32
        ltrasoundGonadLength:
          type: number
        gonadCondition:
          type: string
          maxLength: 50
        uploadedBy:
          type: string
          maxLength: 20
        uploadFilename:
          type: string
          maxLength: 200
    uploadSupplemental:
      type: object
      required:
        - siteId
        - uploadFilename
      properties:
        siteId:
          type: integer
          format: int32
        fFid:
          type: string
          maxLength: 50
        mrFid:
          type: string
          maxLength: 50
        tagnumber:
          type: string
          maxLength: 20
        pitrn:
          type: string
          maxLength: 5
        scuteloc:
          type: string
          maxLength: 1
        scutenum:
          type: number
        scuteloc2:
          type: string
          maxLength: 1
        scutenum2:
          type: number
          maxLength: 1
        elhv:
          type: string
          maxLength: 1
        elcolor:
          type: string
          maxLength: 1
        erhv:
          type: string
          maxLength: 1
        ercolor:
          type: string
          maxLength: 1
        cwtyn:
          type: string
          maxLength: 1
        dangler:
          type: string
          maxLength: 1
        genetic:
          type: string
          maxLength: 1
        geneticsVialNumber:
          type: string
          maxLength: 20
        broodstock:
          type: number
        hatchWild:
          type: number
        speciesId:
          type: integer
          format: int32
        archive:
          type: integer
          format: int32
        head:
          type: number
        snouttomouth:
          type: number
        inter:
          type: number
        outhwidth:
          type: number
        mIb:
          type: number
        lOb:
          type: number
        lIb:
          type: number
        rIb:
          type: number
        rOb:
          type: number
        anal:
          type: number
        orsal:
          type: number
        status:
          type: string
        hatcheryOrigin:
          type: string
          maxLength: 1
        sex:
          type: string
          maxLength: 5
        stage:
          type: string
          maxLength: 5
        recapture:
          type: string
          maxLength: 1
        photo:
          type: string
          maxLength: 1
        geneticNeeds:
          type: string
          maxLength: 100
        otherTagInfo:
          type: string
          maxLength: 500
        comments:
          type: string
          maxLength: 20
        uploadBy:
          type: string
          maxLength: 20
        uploadFilename:
          type: string
          maxLength: 200
    uploadMoriver:
      type: object
      required:
        - siteId
        - uploadFilename
      properties:
        siteId:
          type: integer
          format: int32
        siteFid:
          type: string
          maxLength: 50
        mrFid:
          type: string
          maxLength: 50
        season:
          type: string
          maxLength: 5
        setdate:
          type: string
        subsample:
          type: number
        subsamplepass:
          type: number
        subsamplen:
          type: string
          maxLength: 1
        recorder:
          type: string
          maxLength: 3
        gear:
          type: string
          maxLength: 10
        gearType:
          type: string
          maxLength: 1
        temp:
          type: number
        turbidity:
          type: number
        conductivity:
          type: number
        do:
          type: number
        distance:
          type: number
        width:
          type: number
        netrivermile:
          type: number
        structurenumber:
          type: string
          maxLength: 10
        usgs:
          type: string
          maxLength: 8
        riverstage:
          type: number
        discharge:
          type: number
        u1:
          type: string
          maxLength: 3
        u2:
          type: string
          maxLength: 3
        u3:
          type: string
          maxLength: 4
        u4:
          type: string
          maxLength: 4
        u5:
          type: string
          maxLength: 6
        u6:
          type: string
          maxLength: 6
        u7:
          type: string
          maxLength: 8
        macro:
          type: string
          maxLength: 4
        meso:
          type: string
          maxLength: 5
        habitatrn:
          type: string
          maxLength: 1
        qc:
          type: string
          maxLength: 1
        microStructure:
          type: string
          maxLength: 5
        structureFlow:
          type: string
          maxLength: 1
        structureMod:
          type: string
          maxLength: 1
        setSite_1:
          type: string
          maxLength: 5
        setSite_2:
          type: string
          maxLength: 1
        setSite_3:
          type: string
          maxLength: 1
        startTime:
          type: string
          maxLength: 11
        startLatitude:
          type: number
        startLongitude:
          type: number
        stopTime:
          type: string
          maxLength: 11
        stopLatitude:
          type: number
        stopLongitude:
          type: number
        depth1:
          type: number
        velocitybot1:
          type: number
        velocity08_1:
          type: number
        velocity02or06_1:
          type: number
        depth2:
          type: number
        velocitybot2:
          type: number
        velocity08_2:
          type: number
        velocity02or06_2:
          type: number
        depth3:
          type: number
        velocitybot3:
          type: number
        velocity08_3:
          type: number
        velocity02or06_3:
          type: number
        watervel:
          type: number
        cobble:
          type: number
        organic:
          type: number
        silt:
          type: number
        sand:
          type: number
        gravel:
          type: number
        comments:
          type: string
          maxLength: 2000
        complete:
          type: number
        checkby:
          type: string
          maxLength: 3
        noTurbidity:
          type: string
          maxLength: 1
        noVelocity:
          type: string
          maxLength: 1
        uploadFilename:
          type: string
          maxLength: 200
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string